rename cli/ to cmd/
