--- cli/go.mod
@@ -1,3 +1,0 @@
-module github.com/featurevisor/featurevisor-go/cli
-
-go 1.21
