+++ cmd/commands/test.go
