+++ cmd/go.mod
@@ -1,0 +1,3 @@
+module github.com/featurevisor/featurevisor-go/cmd
+
+go 1.21
