+++ go.work
@@ -2,5 +2,5 @@ go 1.22.0
 
 use (
 	.
-	./cli
+	./cmd
 )
